% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesRegress.R
\name{BayesRegress}
\alias{BayesRegress}
\title{Bayesian Regression}
\usage{
BayesRegress(formula, data)
}
\arguments{
\item{formula}{the formula of the model that is desired to be fit}

\item{data}{the data used for fitting}
}
\value{
regression coefficients as determined by both bayesian and classical regression along with bayesian diagnostic plots/statistics
}
\description{
Bayesian Regression
}
\examples{
BayesRegress(y~x1+x2,data) and BayesRegress(y~x1+x2+x1*x2,data)
}
