% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootStrap.R
\name{Bootstrap}
\alias{Bootstrap}
\title{Bootstrap to Determine Regression Coefficients}
\usage{
Bootstrap(iter, alpha, X, Y)
}
\arguments{
\item{iter}{the number of iterations to perform in bootstrap}

\item{alpha}{the significance level for confidence intervals}

\item{X}{the design matrix}

\item{Y}{the response vector}
}
\value{
regression coefficients and histogram of regression coefficients vs iterations along with comparison to classical regression approach
}
\description{
Bootstrap to Determine Regression Coefficients
}
\examples{
Bootstrap(1000,.05,X,Y) where X and Y are determined by data OR Bootstrap(100000,.1,X,Y)
}
